Remove $schemata, rewrite SPEC and parser
The initial listing of all the names of the schemata in the $schemata
section is redundant; the parser doesn't need to know this information
to check for well-formation, and subsequent analysis would rely on a
constructed object of size linear to the number of schemata in a schema
graph file anyway. Removing $schemata reduces the length and
repetitiveness of Medea schema graph files without affecting our ability
to validate them. The parser logic so far has been rewritten with this
in mind, and the conformance suite has been adjusted accordingly.