Add tests for the 'any schema', modify parser to conform to SPEC.md
According to SPEC.md, a $type specifier is allowed to be missing; in
fact, this is the only way we can specify that any type of JSON value is
acceptable to said schema. However, the parser for Schema required a
$type specifier to be present. This has been corrected.

Currently, the any schema doesn't seem to validate properly, as it
rejects some JSON. Additionally, string validation is inconsistent.