Initial draft of free comonad output
Currently, the validator is very much in the 'validate' camp of Alexis
King's 'parse, not validate' maxim, which is somewhat grating.
It would be much better to return a structured (i.e. annotated) value
demonstrating what precise part of the schema gives rise to its
conformance, on a subtree-by-subtree basis.

My current thinking is that the free comonad is suitable for this, as it
also gives us some useful 'build up' capabilities, which we're currently
rolling from scratch. We should, of course, provide the ability to
'extract' a raw Aeson Value from it, but this also means additional
structural analysis by client code could be done.