Clarify non-circularity requirements for type specifications
The requirement that 'the typing graph be acyclic' is rather
implementation-dependent, and can only be explained by the introduction
of a graph-based model of Medea schemata. Additionally, the
specification layout did not previously allow for any constraints on the
behaviour of a specification (or what is considered valid or not) that
aren't syntactic.

We introduce a new 'Postconditions' section to address the latter issue.
Additionally, we now have a description of the type specification
requirement described in the former case which doesn't rely on a
graph-based model (instead referring to the transitive closure of a
relation), which covers all desired cases.