Clarification, array specification, string specification
The previous version of SPEC.md did not clarify the distinction between
the time when a Medea file is loaded and the time when a Medea
specification is checked against a given JSON file. These are now
clarified, and some statements regarding when error conditions are
indicated are given.

The old array specification has been subdivided into two specifications.
This is based on the two typical use cases of arrays in JSON: either as
tuples (namely, heterogenously-typed collections of fixed size) or as
lists (namely, homogenously-typed collections of varying size). There is
now a specification for each, with their own sub-specifications.

Lastly, strings can now be specified from a finite, non-empty set of
possible values.