Remove inheritance capabilities from schemata, update preconditions
Previously, preconditions for many specifiers permitted a de-facto form
of inheritance. This created a range of issues with validation,
including ambiguities over how this was to be interpreted. As a result,
prerequisites now have much stricter requirements: in particular,
primitive typing requirements are now enforced for any specifiers that
demand them, rather than transitive typing requirements as before.