Carry around our conformance suite
Without this being in the package, we can't have Medea be on Stackage,
as the tests fail to run out of an unpacked sdist. To be honest, we
should probably carry this around _anyway_, because as it currently
stands, people who get Medea off an sdist can't even run the tests.